+++ django_webix/views/generic/list.py
@@ -28,6 +28,7 @@ try:
 except ImportError:
     MultiPolygon = object
 
+
 class WebixListView(WebixBaseMixin,
                     WebixPermissionsMixin,
                     WebixUrlMixin,
@@ -148,7 +149,8 @@ class WebixListView(WebixBaseMixin,
                 except ValueError:
                     print(_('ERROR: geo srid is incorrect'))
                 geo_field = self.model._meta.get_field(geo_field_name)
-                _geo = geo.transform(geo_field.srid, clone=True).buffer(0)  # se l'unione del multipolygon risultasse invalida
+                _geo = geo.transform(geo_field.srid, clone=True).buffer(
+                    0)  # se l'unione del multipolygon risultasse invalida
                 qset = Q(**{geo_field_name + '__within': _geo})
             else:
                 qset = Q()
@@ -243,12 +245,12 @@ class WebixListView(WebixBaseMixin,
 
     def get_fields(self):
         if self.fields is None:
-            return self.fields
+            return None
         else:
             _fields = []
             for _field in self.fields:
                 datalist_column = _field['datalist_column']
-                if type(datalist_column)==dict:
+                if type(datalist_column) == dict:
                     if 'template_string' in datalist_column:
                         template = Template(datalist_column['template'])
                     elif 'template_name' in datalist_column:
@@ -256,10 +258,10 @@ class WebixListView(WebixBaseMixin,
                     else:
                         raise Exception('Template is not defined')
                     context = Context(datalist_column.get('context', {}))
-                else: # string
+                else:  # string
                     template = Template(datalist_column)
                     context = Context({})
-                _field['datalist_column']  = template.render(context)
+                _field['datalist_column'] = template.render(context)
                 _fields.append(_field)
             return _fields
 
@@ -284,7 +286,7 @@ class WebixListView(WebixBaseMixin,
                 sql_filter = ' OR '.join(['({sql})'.format(sql=_sql) for _sql in self.sql_filters])
                 qs = qs.extra(where=[sql_filter])
             # 5. apply django webix filters
-            #raise Exception(self.django_webix_filters)
+            # raise Exception(self.django_webix_filters)
             if self.django_webix_filters is not None:
                 qs = qs.filter(self.get_django_webix_filters_qsets())
 
@@ -308,7 +310,8 @@ class WebixListView(WebixBaseMixin,
                     field_name = field.get('field_name')
                     # TODO: there are no null option
                     _fields_choices[field_name] = list(
-                        self.get_queryset().filter(**{field_name+'__isnull':False}).values_list(field_name, flat=True).distinct().order_by())
+                        self.get_queryset().filter(**{field_name + '__isnull': False}).values_list(field_name,
+                                                                                                   flat=True).distinct().order_by())
 
         return _fields_choices
 
@@ -318,7 +321,7 @@ class WebixListView(WebixBaseMixin,
         if fields is not None:
             for field in fields:
                 if field.get('footer') is not None:
-                    is_footer=True
+                    is_footer = True
         return is_footer
 
     def get_footer(self):
@@ -454,15 +457,15 @@ class WebixListView(WebixBaseMixin,
         _dict_actions = {}
         for _action in _actions:
             _dict_actions[_action.action_key] = {
-                    'func':_action,
-                    'action_key': _action.action_key,
-                    'response_type': _action.response_type,
-                    'allowed_permissions': _action.allowed_permissions,
-                    'short_description': _action.short_description,
-                    'modal_title': _action.modal_title,
-                    'modal_ok': _action.modal_ok,
-                    'modal_cancel': _action.modal_cancel,
-                }
+                'func': _action,
+                'action_key': _action.action_key,
+                'response_type': _action.response_type,
+                'allowed_permissions': _action.allowed_permissions,
+                'short_description': _action.short_description,
+                'modal_title': _action.modal_title,
+                'modal_ok': _action.modal_ok,
+                'modal_cancel': _action.modal_cancel,
+            }
 
         return _dict_actions
 
@@ -520,7 +523,6 @@ class WebixListView(WebixBaseMixin,
         action_name = self.request.POST.get('action', self.request.GET.get('action', None))
         return self.get_actions().get(action_name)['func']
 
-
     def post(self, request, *args, **kwargs):  # all post works like get
         return self.get(request, *args, **kwargs)
 
@@ -552,7 +554,7 @@ class WebixListView(WebixBaseMixin,
         pos = self.get_paginate_start()
         # output must be list and not values of queryset
         data = {
-            "footer": self.get_footer() if pos==0 else None, # footer is computed only for first page
+            "footer": self.get_footer() if pos == 0 else None,  # footer is computed only for first page
             'is_enable_footer': self.is_enable_footer(),
             "count": self.get_paginate_count(),
             "total_count": total_count,
@@ -593,7 +595,8 @@ class WebixListView(WebixBaseMixin,
         self.geo_filter = self.get_geo_filter(self.request)
 
         # 4. SQL FILTERS
-        self.sql_filters = self.get_sql_filters(self.request)  # for now it's only a qsets list # this is shit! but need for old SW (remove for future)
+        self.sql_filters = self.get_sql_filters(
+            self.request)  # for now it's only a qsets list # this is shit! but need for old SW (remove for future)
 
         # 5. DJANGO WEBIX FILTERS
         self.django_webix_filters = self.get_django_webix_filters(self.request)
@@ -619,7 +622,7 @@ class WebixListView(WebixBaseMixin,
             'orders': self.get_ordering(),
             'actions': self.get_actions(),
             'choices_filters': self.get_choices_filters(),
-            'footer': self.get_footer() if not self.enable_json_loading else None, # footer only if not paging
+            'footer': self.get_footer() if not self.enable_json_loading else None,  # footer only if not paging
             'is_enable_footer': self.is_enable_footer(),
             'get_pk_field': self.get_pk_field(),
             'objects_datatable': self.get_objects_datatable(),
@@ -637,11 +640,12 @@ class WebixListView(WebixBaseMixin,
             'paginate_start_key': self.paginate_start_key,
             # extra filters
             'is_installed_django_webix_filter': self.is_installed_django_webix_filter(),
-            'qsets_filters': self.get_qsets_filters_str(self.request), # for init when template is loaded
-            'qsets_locked_filters': self.get_qsets_locked_filters_str(self.request), # for init when template is loaded
-            'geo_filter': self.get_geo_filter_str(self.request), # for init when template is loaded
-            'sql_filters': self.get_sql_filters_str(self.request), # for init when template is loaded
-            'django_webix_filters': self.get_django_webix_filters(self.request), # for init when template is loaded
+            'qsets_filters': self.get_qsets_filters_str(self.request),  # for init when template is loaded
+            'qsets_locked_filters': self.get_qsets_locked_filters_str(self.request),
+            # for init when template is loaded
+            'geo_filter': self.get_geo_filter_str(self.request),  # for init when template is loaded
+            'sql_filters': self.get_sql_filters_str(self.request),  # for init when template is loaded
+            'django_webix_filters': self.get_django_webix_filters(self.request),  # for init when template is loaded
         })
         return context